pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-orchestrator"
version = "0.1.0"
description = "Multi-AI Code Orchestration System for automated code generation and review"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Auto Coder Multi Agents" }
]
keywords = [
    "ai",
    "code-generation",
    "automation",
    "orchestration",
    "claude",
    "llm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
]
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "json-repair>=0.28.0",
    "pybreaker>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
ai-orchestrator = "ai_orchestrator.__main__:cli_main"

[project.urls]
Homepage = "https://github.com/auto-coder-multi-agents/ai-orchestrator"
Repository = "https://github.com/auto-coder-multi-agents/ai-orchestrator"

[tool.hatch.build.targets.wheel]
packages = ["ai_orchestrator"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["ai_orchestrator"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["pybreaker.*", "json_repair.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["ai_orchestrator/tests"]
addopts = "-v --tb=short"
